refactor(evals): extract linkQualityEvaluator into its own module
Moves the link-quality scoring logic out of documentation.eval.ts into
evals/evaluators/linkQualityEvaluator.ts so it can be reused across evals
and read independently from any specific experiment. Exposes a
createLinkQualityEvaluator factory so consumers can override the
allowed-prefix list or the evaluator name.
